Foxy build fixes (and a warning fix)

+++ tesseract_planning_server/src/tesseract_planning_server_node.cpp
@@ -48,7 +48,7 @@ int main(int argc, char** argv)
 
   // Defaults
   bool publish_environment{ false };
-  int cache_size{ 5 };
+  long cache_size{ 5 };
   double cache_refresh_rate{ 0.1 };
 
   std::string monitor_namespace = node->declare_parameter("monitor_namespace", "");
